+++ include/avian/util/fixed-allocator.h
@@ -24,13 +24,13 @@ class FixedAllocator : public Allocator {
  public:
   FixedAllocator(Aborter* a, Slice<uint8_t> memory);
 
-  virtual void* tryAllocate(unsigned size);
+  virtual void* tryAllocate(size_t size);
 
-  void* allocate(unsigned size, unsigned padAlignment);
+  void* allocate(size_t size, unsigned padAlignment);
 
-  virtual void* allocate(unsigned size);
+  virtual void* allocate(size_t size);
 
-  virtual void free(const void* p, unsigned size);
+  virtual void free(const void* p, size_t size);
 
   Aborter* a;
   Slice<uint8_t> memory;
