+++ src/main.cpp
@@ -68,12 +68,12 @@ const char* mainClass(const char* jar)
     {
     }
 
-    virtual void* tryAllocate(unsigned size)
+    virtual void* tryAllocate(size_t size)
     {
       return s->tryAllocate(size);
     }
 
-    virtual void* allocate(unsigned size)
+    virtual void* allocate(size_t size)
     {
       void* p = tryAllocate(size);
       if (p == 0) {
@@ -82,7 +82,7 @@ const char* mainClass(const char* jar)
       return p;
     }
 
-    virtual void free(const void* p, unsigned)
+    virtual void free(const void* p, size_t)
     {
       s->free(p);
     }
